I got problems
ok so after a lot of research i came up with this program that god knows
how
it works(i understand a little XD).
so anyway one minute it kinda worked for me but then i changed the item
price to int instead of double, and since then it wont find the shop.txt
even though it exists.
note that instead of rewriting the file each time, or changing it( which
i havent realized how to do) i rewrote it in the end so that while in
the loop all changes are virtual.
also this is kinda my first time using "try and catch" so i think that
"exception e" is any exception but i might be wrong. please correct me
if i am.