V2 of the program not finished *need help*
So i did kinda everything you said( if i missed something please tell
me).
i got a problem that i cant solve and i need your help with it.
I described it in the .md file so make sure to read it.
also the print option for the list int ShopItemList is basic because
when I
wrote the printing code for some reason it skiped the first value of the
list.
oh and another very importent thing: be sure to delete the content of
the shop.txt after each  time you run the program. because they way it
wrotes it makes the reading part to malfunction.