File creation finished (I hope :P)
ok so i changed the program how you told me to and it works!
:D
so im gonna start the graphic project later because my sister (again)
wants the computer.
btw i talked to gal and he says he think it tould be just great to learn
programming. so maybe he will do what he did to you and wont call (or in
this case email) back, but for the time being consider him a part of the
classes. (like in school not in java XD).