Why jave hates me so much?
ok so i have 3 problems
1) still in the firts task it wont do a thing when i press a
the next 2 are in the second task(random falling shape)
1) how do i create a random shape without using a polygon? i managed to
create a basic rectangle but nothing nore.
2) for some reason when i call tick it does something wierd. the window
is 99% black and thats it. nothing is happening and i cant close the
window unless i hit the stop button in eclipse. it does it also when i
try to move the rectangle in the x axis. i guess it because some kind of
error but what error could that be?